# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  context.Response:
    properties:
      data: {}
      err:
        type: string
      msg:
        type: string
      success:
        type: boolean
      trace_id:
        type: string
    type: object
...
  model.ListRes:
    properties:
      total:
        type: integer
    type: object
...
  model.User:
    properties:
      avatar:
        type: string
      builtin:
        type: boolean
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      intro:
        type: string
      invisible:
        type: boolean
      key:
        type: string
      last_login:
        type: integer
      name:
        type: string
      org_ids:
        items:
          type: integer
        type: array
      password:
        type: string
      point:
        type: integer
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: integer
      web_notify:
        type: boolean
    type: object
  model.UserInfo:
    properties:
      auth_type:
        type: integer
      avatar:
        type: string
      builtin:
        type: boolean
      cors:
        type: boolean
      email:
        type: string
      intro:
        type: string
      key:
        type: string
      no_password:
        type: boolean
      org_ids:
        items:
          type: integer
        type: array
      point:
        type: integer
      role:
        $ref: '#/definitions/model.UserRole'
      uid:
        type: integer
      username:
        type: string
      web_notify:
        type: boolean
    type: object
...
  model.UserRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - UserRoleUnknown
    - UserRoleAdmin
    - UserRoleOperator
    - UserRoleUser
    - UserRoleGuest
    - UserRoleMax
...
  svc.UserListItem:
    properties:
      avatar:
        type: string
      builtin:
        type: boolean
      created_at:
        type: integer
      email:
        type: string
      id:
        type: integer
      last_login:
        type: integer
      name:
        type: string
      org_ids:
        items:
          type: integer
        type: array
      org_names:
        items:
          type: string
        type: array
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: integer
    type: object
  svc.UserLoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  svc.UserRegisterReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  svc.UserStatisticsRes:
    properties:
      answer_count:
        type: integer
      avatar:
        type: string
      blog_count:
        type: integer
      intro:
        type: string
      name:
        type: string
      point:
        type: integer
      qa_count:
        type: integer
      role:
        $ref: '#/definitions/model.UserRole'
    type: object
  svc.UserUpdateReq:
    properties:
      email:
        type: string
      name:
        type: string
      org_ids:
        items:
          type: integer
        type: array
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        maximum: 4
        minimum: 1
    type: object
...
info:
  contact: {}
  description: KoalaQA API swagger document.
  title: KoalaQA API
  version: "2.0"
paths:
...
  /admin/user:
    get:
      parameters:
      - in: query
        name: email
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: org_id
        type: integer
      - in: query
        name: org_name
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        in: query
        name: role
        type: integer
        x-enum-varnames:
        - UserRoleUnknown
        - UserRoleAdmin
        - UserRoleOperator
        - UserRoleUser
        - UserRoleGuest
        - UserRoleMax
      - in: query
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.ListRes'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/svc.UserListItem'
                        type: array
                    type: object
              type: object
      summary: list user
      tags:
      - user
  /admin/user/{user_id}:
    delete:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: delete user
      tags:
      - user
    get:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      summary: user detail
      tags:
      - user
    put:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/svc.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: update user
      tags:
      - user
...
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserInfo'
              type: object
      summary: user detail
      tags:
      - user
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: avatar
        in: formData
        name: avatar
        type: file
      - in: formData
        name: email
        type: string
      - in: formData
        name: intro
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: old_password
        type: string
      - in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: update user
      tags:
      - user
  /user/{user_id}:
    get:
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  $ref: '#/definitions/svc.UserStatisticsRes'
              type: object
      summary: stat user info
      tags:
      - user
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/svc.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: user login
      tags:
      - user
  /user/login/cors:
    post:
      consumes:
      - application/json
      parameters:
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/svc.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: user cors login
      tags:
      - user
  /user/login/third:
    get:
      parameters:
      - in: query
        name: app
        type: boolean
      - in: query
        name: redirect
        type: string
      - in: query
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/context.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: get user third login url
      tags:
      - user
...
  /user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: request params
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/svc.UserRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/context.Response'
      summary: register user
      tags:
      - user
...
swagger: "2.0"